Move filling of SRGlobal to FillTrue.cxx
