+++ sbncode/CAFMaker/CAFMaker_module.cc
@@ -337,13 +337,6 @@ void CAFMaker::beginJob() {
 }
 
 //......................................................................
-void CopyTMatrixDToVector(const TMatrixD& m, std::vector<float>& v)
-{
-  const double& start = m(0, 0);
-  v.insert(v.end(), &start, &start + m.GetNoElements());
-}
-
-//......................................................................
 void CAFMaker::beginRun(art::Run& run) {
   // fDetID = geom->DetId();
   fDet = (Det_t)1;//(Det_t)fDetID;
@@ -380,27 +373,7 @@ void CAFMaker::beginRun(art::Run& run) {
     fPrevWeightPSet[label] = *wgt_params;
 
     for(const sbn::evwgh::EventWeightParameterSet& pset: *wgt_params){
-      SRWeightPSet& cafpset = global.wgts.emplace_back();
-
-      cafpset.name = pset.fName;
-      cafpset.type = caf::ReweightType_t(pset.fRWType);
-      cafpset.nuniv = pset.fNuniverses;
-      if(pset.fCovarianceMatrix) CopyTMatrixDToVector(*pset.fCovarianceMatrix, cafpset.covmx);
-
-      fWeightPSetIndex[cafpset.name] = global.wgts.size()-1;
-
-      for(const auto& it: pset.fParameterMap){
-        const sbn::evwgh::EventWeightParameter& param = it.first;
-        const std::vector<float>& vals = it.second;
-
-        SRWeightParam cafparam;
-        cafparam.name = param.fName;
-        cafparam.mean = param.fMean;
-        cafparam.width = param.fWidth;
-        cafparam.covidx = param.fCovIndex;
-
-        cafpset.map.emplace_back(cafparam, vals);
-      } // end for it
+      FillSRGlobal(pset, global, fWeightPSetIndex);
     } // end for pset
   } // end for label
